line_ai_bot/requirements.txt
aenum==3.1.15 
aiohappyeyeballs==2.4.4 
aiohttp==3.11.11 
aiosignal==1.3.2 
annotated-types==0.7.0 
anyio==4.8.0 
attrs==24.3.0 
blinker==1.9.0 
certifi==2024.12.14 
charset-normalizer==3.4.1 
click==8.1.8 
colorama==0.4.6 
deprecated==1.2.15 
distro==1.9.0 
flask==3.1.0 
frozenlist==1.5.0 
future==1.0.0 
gunicorn==23.0.0 
h11==0.14.0 
httpcore==1.0.7 
httpx==0.28.1 
idna==3.10 
itsdangerous==2.2.0 
jinja2==3.1.5 
jiter==0.8.2 
line-bot-sdk==3.14.2 
markupsafe==3.0.2 
multidict==6.1.0 
openai==1.59.3 
packaging==24.2 
propcache==0.2.1 
pydantic-core==2.27.2 
pydantic==2.10.4 
python-dateutil==2.9.0.post0 
requests==2.32.3 
six==1.17.0 
sniffio==1.3.1 
tqdm==4.67.1 
typing-extensions==4.12.2 
urllib3==2.3.0 
werkzeug==3.1.3 
wrapt==1.17.0 
yarl==1.18.3 
